feat(View): move interactor style up to View
The View component controls the interactor style, as that's the primary
component for working with multiple views.

Other changes:
- Fixes the interactor style being deleted before it gets removed from
the interactor.
